master
This commit updates workspaces.

Before this commit anytime a workspace was changed the current state
of that workspace was saved. Any return to that workspace would return
to that last state before leaving the workspace.

This is the right behavior when changing workspaces. When moving to
the same workspace, though it should snap back to the last saved state
at that workspace and not save the current one.

Moving to the same workspace that one is on is a way of discarding the
current state as not important.